Refactor - Minor changes
